refactor: evaluation 관련 pk를 BIGINT로 변경 & 이전 평가 불러오기 정상화

+++ src/main/java/com/kustaurant/kustaurant/evaluation/evaluation/infrastructure/EvaluationSituationId.java
@@ -8,13 +8,13 @@ import java.util.Objects;
 @Getter
 @Setter
 public class EvaluationSituationId implements Serializable {
-    private Integer evaluation;
-    private Integer situation;
+    private Long evaluation;
+    private Long situation;
 
     public EvaluationSituationId() {
     }
 
-    public EvaluationSituationId(Integer evaluation, Integer situation) {
+    public EvaluationSituationId(Long evaluation, Long situation) {
         this.evaluation = evaluation;
         this.situation = situation;
     }
