+++ src/main/java/com/kustaurant/kustaurant/evaluation/evaluation/infrastructure/situation/SituationRepository.java
@@ -4,8 +4,8 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface SituationRepository extends JpaRepository<SituationEntity,Integer> {
+public interface SituationRepository extends JpaRepository<SituationEntity,Long> {
     SituationEntity findBySituationName(String situationName);
 
-    Optional<SituationEntity> findBySituationId(Integer situationId);
+    Optional<SituationEntity> findBySituationId(Long situationId);
 }
